requirements.txt
Pillow==8.2.0
pyfiglet==0.8.post1
